chore: web content system action supporting PUT,DELETE http verbs
